Add early support for installing front- and back-end tookit builds into the same prefix.  Hpclink uses tools on the front end (fnbounds) and libraries for the back end, so it needs to have both in the same tree.
The solution here is to rename the launch scripts with a back-end
suffix and to rename the pkg-lib and pkg-libexec directories.

Note: this is still in flux.  Currently, the --enable-back-end=LABEL
option both sets the back-end label and selects a subset of back-end
tools to build.  For now, we only build hpcrun and fnbounds for the
back end.

Also, this patch only installs the files under different names.  The
hpcrun and hpclink scripts don't yet run correctly on the back end.