Partitioning x86ISA class into (1) the default implementation for all architectures using binutils (x86ISA.cpp), and (2) x86 default implementation using XED (x86ISA_xed.cpp). All generic methods that are used by Binutils and XED implementation are defined in x86ISA_helper.cpp (new file)
Attention: this commit will enable XED as the default x86 (32 and 64) decoding in hpcstruct 

So far, the result of hpcstruct with XED and binutils are almost similar, in exception when decoding a MIC backend program (a program that is built using specialized MIC flags such as -mmic). 
In general, the tree is not different, but the number of instructions in a basic block can be different. It's hard to tell which one is the correct one.